+++ ast_parcours.h
@@ -2,7 +2,7 @@
 void afficherA(Ast expr) ;
 // affiche l'arbre abstrait de l'expression arithmetique expr
 
-int evaluation(Ast expr) ;
+double evaluation(Ast expr) ;
 // calcule la valeur de l'expression arithmetique expr
 // FONCTION A COMPLETER !
